🔒 fix: use timing-safe comparison for webhook authentication
